Added option to delete submission. Added custom alert-boxes for that.

+++ BSS/src/components/SubmittedAssignments.vue
@@ -4,10 +4,13 @@
       <h1>Oddane naloge</h1>
     </div>
     <div class="assignment">
-      <div class="submitted-assignment" v-for ="a in assignment" @click="assignmentInfo(a.id)">
-        <h2>{{a.tittle}}</h2>
-        <p>{{a.subject}}</p>
-        <p>{{a.deadline}}</p>
+      <div class="submitted-assignment" v-for ="a in assignment">
+        <div @click="assignmentInfo(a.id)" class="assignment-inner-div">
+          <h2>{{a.tittle}}</h2>
+          <p>{{a.subject}}</p>
+          <p>{{a.deadline}}</p>
+        </div>
+        <p><img src="../assets/delete-icon.png" alt="Izbriši" @click="deleteSubmission(a.file)"></p>
       </div>
     </div>
   </div>
@@ -15,6 +18,7 @@
 
 <script lang="ts">
 import axios from "axios";
+import Swal from "sweetalert2";
 
 export default {
   data() {
@@ -37,9 +41,19 @@ export default {
                   this.assignment.push(response.data.assignments[i]);
                 }
               } else if (response.data.error == "session") {
-                alert("Session has expired, please log in again!");
-                sessionStorage.clear();
-                localStorage.clear();
+                Swal.fire({
+                  title: "Ponovna prijava",
+                  text: "Šment, vaša seja se je iztekla! Ponovno se prijavite da nadaljujete z delom.",
+                  icon: "warning",
+                  confirmButtonText: "Na prijavno stran.",
+                  confirmButtonColor: "#4377df"
+                })
+                    .then(function (isConfirmed){
+                      if (isConfirmed.isConfirmed || isConfirmed.isDismissed){
+                        sessionStorage.clear();
+                        localStorage.clear();
+                      }
+                    })
                 this.$router.push('/');
               }
             })
@@ -49,6 +63,65 @@ export default {
     assignmentInfo(id: string){
       sessionStorage.setItem('assignmentId', id);
       this.$router.push('/assignments');
+    },
+    deleteSubmission(file: string){
+      const token = sessionStorage.getItem('token');
+      const subjectId = sessionStorage.getItem('subjectId');
+      if (token != null && subjectId != null){
+        /*define the API URL*/
+        const path = "https://smv.usdd.company/API/public/api/"
+        /*Create new FormData instance and append the needed data*/
+        const data = new FormData();
+        data.append('token', token);
+        data.append('subjectId', subjectId);
+        data.append('fileName', file)
+
+        /*show an alert-box that prevents user from accidentally deleting a submission*/
+
+        Swal.fire({
+          title: "Odstrani oddajo?",
+          text: "Oddaja bo odstranjena iz sistema, datoteka pa izbrisana. Tega dejanja ni mogoče razveljaviti. Želite nadaljevati?",
+          icon: "question",
+          confirmButtonText: 'Da, odstrani oddajo',
+          confirmButtonColor: "#e63946",
+          showDenyButton: true,
+          denyButtonText: 'Ne, zadovoljen sem s trenutno oddajo.',
+          denyButtonColor: '#4377df'
+        })
+            .then(function (isSuccess){
+              if(isSuccess.isConfirmed){
+                axios.post(path + 'assignment/deleteSubmission', data)
+                    .then((response) => {
+                      if (response.data.success == 'true'){
+                        Swal.fire({
+                          title: "Oddaja odstranjena",
+                          text: "Uspešno ste odstranili oddajo naloge.",
+                          icon: "success",
+                          confirmButtonText: "Razumem",
+                          confirmButtonColor: "#4377df"
+                        })
+                      } else {
+                        Swal.fire({
+                          title: "Napaka",
+                          text: "Ojoj, prišlo je do nepričakovane napake! Poskusite ponovno ali kontaktirajte administratorja.",
+                          icon: "error",
+                          confirmButtonText: "Ti šment!",
+                          confirmButtonColor: "#4377df"
+                        })
+                      }
+                    })
+              } else if (isSuccess.isDenied){
+                Swal.fire({
+                  title: "Oddaja ohranjena",
+                  text: "Brez skrbi! Vaša prejšna oddaja je na varnem.",
+                  icon: "success",
+                  confirmButtonText: "Fjuu, čudovito.",
+                  confirmButtonColor: "#4377df"
+                })
+              }
+            })
+        this.$forceUpdate();
+      }
     }
   },
   created() {
@@ -79,9 +152,14 @@ export default {
   border-bottom: 4px solid #2e5baa;
   padding-bottom: 2vh;
   padding-top: 1vh;
+}
+.assignment-inner-div{
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: left;
 
 }
-
 .submitted-assignment:hover {
   border-bottom: 4px solid #5891d3;
   transition: 0.2s ease-in-out;
@@ -92,6 +170,9 @@ export default {
   font-size: large;
   width: 15vw;
   line-break: auto;
+}
+.submitted-assignment img{
+  width: 3vw;
 }
 .submitted-assignment h2,p {
   padding-top: 1vh;
